feat: sections feedbacks and protov4 feedback
