+++ src/app/ui/pages/studies/components/import-study/import-study.component.html
@@ -1,0 +1,55 @@
+<div class="flex flex-col">
+  <div class="flex justify-center mt-10">
+    <div
+      (click)="file.click()"
+      (keydown.enter)="file.click()"
+      (keydown.space)="file.click()"
+      tabindex="0"
+      class="flex flex-col gap-4 border-2 border-gray-200 rounded-md p-4 bg-gray-50 w-[18rem] h-[12rem] items-center justify-center border-dashed cursor-pointer"
+    >
+      <div class="flex flex-col gap-2 text-center items-center">
+        <app-icon icon="upload" />
+        <label for="study-file" class="text-sm text-gray-500" i18n>
+          Upload one or multiple protoV4 files
+        </label>
+      </div>
+      <input
+        id="study-file"
+        form="file"
+        hidden
+        #file
+        multiple
+        name="file"
+        type="file"
+        onclick="this.value = null"
+        (change)="loadProtoV4File($event)"
+      />
+    </div>
+  </div>
+  @if (this.newStudies().length > 0) {
+    <div class="flex justify-center mt-4">
+      <ul i18n class="flex flex-col gap-4 w-[40rem]">
+        Imported studies:
+        @for (study of this.newStudies(); track study) {
+          <li class="flex flex-col gap-2">
+            <div class="items-center flex justify-between">
+              <div class="text-lg items-center flex gap-2">
+                <app-icon icon="check" class="text-green-500" />
+                {{ study.title }}
+              </div>
+              <button
+                app-btn
+                btnSize="s"
+                btnStyle="text"
+                (click)="deleteStudy(study.uuid)"
+              >
+                <app-icon icon="delete" class="text-red-500" />
+              </button>
+            </div>
+            <p-divider class="!my-1" />
+          </li>
+        }
+      </ul>
+    </div>
+  }
+</div>
