+++ src/app/ui/pages/studies/components/new-study-modal/new-study-modal.component.ts
@@ -21,11 +21,6 @@ import { IconComponent } from '@src/app/ui/shared/components/atoms/icon/icon.com
 import { ButtonComponent } from '@src/app/ui/shared/components/atoms/button/button.component';
 import { Router } from '@angular/router';
 import { FileUploadModule } from 'primeng/fileupload';
-import Papa from 'papaparse';
-import {
-  ProtoV4Parameters,
-  ProtoV4Support
-} from '@src/app/core/data/database/interfaces/protoV4';
 
 const newStudy = (): StudyModel => {
   return {
@@ -72,9 +67,6 @@ export class NewStudyModalComponent {
   titleLength = computed(() => this.title().length ?? 0);
   descriptionLength = computed(() => this.description().length ?? 0);
   loading = signal<boolean>(false);
-  studyFromProtoV4 = signal<Pick<StudyModel, 'sections' | 'shareable'> | null>(
-    null
-  );
 
   updateTitle(title: string) {
     this.title.set(title);
@@ -82,78 +74,6 @@ export class NewStudyModalComponent {
 
   updateDescription(description: string) {
     this.description.set(description);
-  }
-
-  loadProtoV4File(event: Event) {
-    this.loading.set(true);
-    const file = (event.target as HTMLInputElement).files?.item(0);
-
-    const convertToNumber = (value: string) => {
-      return Number(value.replace(',', '.'));
-    };
-    if (file) {
-      const reader = new FileReader();
-      reader.onload = (e) => {
-        const rawParameters: string[] = [];
-        const result = e.target?.result as string;
-        if (!result) {
-          this.loading.set(false);
-          return;
-        }
-        const csvSupports = result
-          .split('\n')
-          .map((line: string) => {
-            const parts = line.split(';');
-            rawParameters.push(parts.pop()?.replace('\r', '') ?? '');
-            parts.pop();
-            return parts.join(';');
-          })
-          .join('\n');
-        const parameters: ProtoV4Parameters = {
-          conductor: rawParameters[3],
-          cable_amount: convertToNumber(rawParameters[5]),
-          temperature_reference: convertToNumber(rawParameters[7]),
-          parameter: convertToNumber(rawParameters[9]),
-          cra: convertToNumber(rawParameters[11]),
-          temp_load: convertToNumber(rawParameters[13]),
-          wind_load: convertToNumber(rawParameters[15]),
-          frost_load: convertToNumber(rawParameters[17]),
-          project_name: rawParameters[19]
-        };
-        Papa.parse(csvSupports as string, {
-          header: true,
-          skipEmptyLines: true,
-          complete: (jsonResults: Papa.ParseResult<Record<string, string>>) => {
-            const supports: ProtoV4Support[] = jsonResults.data.map(
-              (support: Record<string, string>) => {
-                return {
-                  ...support,
-                  nom: support.nom,
-                  num: convertToNumber(support.num),
-                  portée: convertToNumber(support.portée),
-                  angle_ligne: convertToNumber(support.angle_ligne),
-                  ctr_poids: convertToNumber(support.ctr_poids),
-                  long_bras: convertToNumber(support.long_bras),
-                  long_ch: convertToNumber(support.long_ch),
-                  pds_ch: convertToNumber(support.pds_ch),
-                  surf_ch: convertToNumber(support.surf_ch),
-                  alt_acc: convertToNumber(support.alt_acc),
-                  suspension: support.suspension === 'FAUX' ? false : true,
-                  ch_en_V: support.ch_en_V === 'FAUX' ? false : true
-                };
-              }
-            );
-            const study = this.studiesService.createStudyFromProtoV4(
-              supports,
-              parameters
-            );
-            this.studyFromProtoV4.set(study);
-            this.loading.set(false);
-          }
-        });
-      };
-      reader.readAsText(file);
-    }
   }
 
   constructor(
@@ -172,7 +92,7 @@ export class NewStudyModalComponent {
   async onSubmit() {
     if (this.mode() === 'new') {
       const uuid = await this.studiesService.createStudy({
-        ...(this.studyFromProtoV4() ?? newStudy()),
+        ...newStudy(),
         title: this.title(),
         description: this.description()
       });
