+++ src/app/ui/pages/studies/studies.component.html
@@ -29,6 +29,8 @@
       />
     </p-tabpanel>
     <p-tabpanel value="searchStudy"> </p-tabpanel>
-    <p-tabpanel value="importStudy"> </p-tabpanel>
+    <p-tabpanel value="importStudy">
+      <app-import-study />
+    </p-tabpanel>
   </p-tabpanels>
 </p-tabs>
