+++ src/app/ui/pages/study/tabs/sections/sectionsTab.component.html
@@ -74,7 +74,7 @@
             (click)="
               openInitialConditionModal(
                 section,
-                createInitialCondition(),
+                createInitialCondition(section),
                 'create'
               )
             "
@@ -234,7 +234,7 @@
               (click)="
                 openInitialConditionModal(
                   section,
-                  createInitialCondition(),
+                  createInitialCondition(section),
                   'create'
                 )
               "
